fix: fix service name
