+++ src/app/services/cart-service/current-cart/index.ts
@@ -1,0 +1,1 @@
+export * from './current-cart';
