--- src/app/services/cart-service/currentCart/index.ts
@@ -1,1 +1,0 @@
-export * from './current-cart';
